requirements-ci.txt
alembic==1.17.1
altair==5.5.0
aniso8601==10.0.1
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
apispec==6.6.1
apispec-webframeworks==1.1.0
arrow==1.4.0
attrs==25.4.0
Automat==25.4.16
bidict==0.23.1
binaryornot==0.4.4
blinker==1.9.0
boto3==1.34.113
botocore==1.34.162
cachetools==6.2.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.3.0
cloudpickle==3.1.1
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.3
cookiecutter==2.6.0
cryptography==46.0.3
cycler==0.12.1
databricks-sdk==0.70.0
deepdiff==7.0.1
dnspython==2.8.0
docker==7.1.0
et_xmlfile==2.0.0
fastapi==0.120.2
filelock==3.20.0
Flask==3.1.2
flask-cors==5.0.1
Flask-RESTful==0.3.10
Flask-SocketIO==5.4.1
fonttools==4.60.1
fsspec==2025.9.0
gevent==24.11.1
gevent-websocket==0.10.1
gitdb==4.0.12
gitignore_parser==0.1.11
GitPython==3.1.45
google-auth==2.42.0
graphene==3.4.3
graphql-core==3.2.6
graphql-relay==3.2.0
greenlet==3.2.4
h11==0.16.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.7.0
incremental==24.7.2
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
kthread==0.2.3
Mako==1.3.10
Markdown==3.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.21.2
matplotlib==3.10.7
mdurl==0.1.2
mlflow==3.5.1
mlflow-skinny==3.5.1
mlflow-tracing==3.5.1
mpmath==1.3.0
narwhals==2.7.0
networkx==3.3
numpy==2.3.3
openpyxl==3.1.2
opentelemetry-api==1.38.0
opentelemetry-proto==1.38.0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
ordered-set==4.1.0
packaging==25.0
pandas==2.2.2
passlib==1.7.4
pillow==11.3.0
pluggy==1.6.0
protobuf==6.32.1
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
Pygments==2.19.2
pymongo==4.7.2
pyparsing==3.2.5
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.12.3
python-slugify==8.0.4
python-socketio==5.14.2
pytz==2024.1
pywin32==311
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.10.4
scikit-learn==1.7.2
scipy==1.16.2
simple-websocket==1.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.30
sqlparse==0.5.3
starlette==0.49.1
streamlit==1.50.0
sympy==1.13.1
taipy==4.1.0
taipy-common==4.1.0
taipy-core==4.1.0
taipy-gui==4.1.0
taipy-rest==4.1.0
taipy-templates==4.1.0
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.5.2
tqdm==4.67.1
Twisted==24.7.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
tzlocal==5.2
urllib3==2.5.0
uvicorn==0.38.0
waitress==3.0.2
watchdog==4.0.1
Werkzeug==3.1.3
wsproto==1.2.0
zipp==3.23.0
zope.event==6.0
zope.interface==8.0.1

